feat: preserve open file when saving template
