+++ templates/frontend/src/components/pages/create-template/create-template.jsx
@@ -1,27 +1,6 @@
 import { useRouter } from 'next/router';
 
-import { gql, useGqlClient } from 'api/graphql';
 import TemplateForm from 'components/shared/template-form';
-
-const createTemplateQuery = gql`
-  mutation createTemplate(
-    $name: String!
-    $prompts: String
-    $files: String!
-    $templateGroupId: String
-  ) {
-    insert_template(
-      object: { name: $name, files: $files, prompts: $prompts, template_group_id: $templateGroupId }
-    ) {
-      id
-      name
-      prompts
-      files
-      template_group_id
-      owner_id
-    }
-  }
-`;
 
 const defaultTemplateValues = {
   name: '',
@@ -30,32 +9,12 @@ const defaultTemplateValues = {
 };
 
 const CreateTemplate = () => {
-  const gqlClient = useGqlClient();
-
   const router = useRouter();
 
   const { groupId } = router.query;
 
-  const handleSave = async ({ name, prompts, files, templateGroupId }) => {
-    const res = await gqlClient.request(createTemplateQuery, {
-      name,
-      prompts,
-      files,
-      ...(templateGroupId ? { templateGroupId } : {}),
-    });
-
-    const templateId = res?.insert_template?.id || null;
-
-    if (templateId) {
-      router.push(`/template/${templateId}/edit`);
-    }
-  };
   return (
-    <TemplateForm
-      initialValues={{ ...defaultTemplateValues, groupId }}
-      isCreatingNewTemplate
-      onSave={handleSave}
-    />
+    <TemplateForm initialValues={{ ...defaultTemplateValues, groupId }} isCreatingNewTemplate />
   );
 };
